Split DeckSettings/DeckGroup.py into DeckGroupParts/
